+++ spec/base/behaviour_spec.rb
@@ -32,59 +32,8 @@ RSpec.describe Employee::Collection do
     end
   end
 
-  describe '#uniform_collection_attribute' do
-    let(:collection_class){ described_class }
-    describe 'boolean attribute' do
-      let(:record) { Struct.new(:admin) }
-
-      it 'returns false for mixed falsy boolean values' do
-        expect( collection_class.new([record.new, record.new(false)]).uniform_collection_attribute(:admin)).to be false
-      end
-
-      it 'returns nil for mixes boolean values' do
-        expect( collection_class.new([record.new, record.new(true)]).uniform_collection_attribute(:admin)).to be nil
-      end
-
-      it 'returns true for all truthy values' do
-        expect( collection_class.new([record.new(true), record.new(true)]).uniform_collection_attribute(:admin)).to be true
-      end
-    end
-
-    describe 'untyped attribute' do
-      let(:record) { Struct.new(:section) }
-      it 'returns nil for mixed values truthy and falsy' do
-        collection = collection_class.new([record.new, record.new('NT2')])
-        expect( collection.uniform_collection_attribute :section ).to be nil
-      end
-
-      it 'returns nil for mixed truthy values' do
-        collection = collection_class.new([record.new('NT1'), record.new('NT2')])
-        expect( collection.uniform_collection_attribute :section ).to be nil
-      end
-
-      it 'returns the value for all the same values, but does not set the value' do
-        collection = collection_class.new([record.new('NT2'), record.new('NT2')])
-        expect( collection.uniform_collection_attribute :section ).to eq 'NT2'
-        expect( collection.section ).to be nil
-      end
-
-      it 'returns the value for all the same values, and sets the value if set_if_nil is given' do
-        collection = collection_class.new([record.new('NT2'), record.new('NT2')])
-        expect( collection.uniform_collection_attribute :section, set_if_nil: true ).to eq 'NT2'
-        expect( collection.section ).to eq 'NT2'
-      end
-
-      it 'returns the value for all the same values, and does not set the value if set_if_nil is given, but already set (eg: by invalid form)' do
-        collection = collection_class.new([record.new('NT2'), record.new('NT2')], section: 'Invalid set form value')
-        expect( collection.uniform_collection_attribute :section, set_if_nil: true ).to eq 'NT2'
-        expect( collection.section ).to eq 'Invalid set form value'
-      end
-    end
-  end
-
   describe "array like behaviour" do
     collection = described_class.new([1])
     Array.wrap( collection ).should == collection
   end
 end
-