+++ spec/dummy/app/models/project/collection.rb
@@ -1,4 +1,5 @@
 class Project::Collection < RecordCollection::Base
   attribute :finished, type: Boolean
+  attribute :start_date, type: Date
   attribute :description
 end