+++ spec/rails/form_builder_spec.rb
@@ -1,8 +1,10 @@
 require 'spec_helper'
 
 describe ActionView::Helpers::FormBuilder do
-  let(:employee){ Employee.create section: 'SE1' }
-  let(:arguments){ {form_object: Employee::Collection.new([employee])} }
+  let(:employee_1){ Employee.create section: 'SE1' }
+  let(:employee_2){ Employee.create section: 'SE1' }
+  let(:collection_class) { Employee::Collection }
+  let(:arguments){ {form_object: collection_class.new([employee_1, employee_2])} }
   subject{ described_class.new :collection, arguments[:form_object], @template, {}}
   before do
     # http://pivotallabs.com/testing-custom-form-builder/
@@ -14,13 +16,11 @@ describe ActionView::Helpers::FormBuilder do
 
   describe '.collection_ids' do
     it 'returns the collection ids as hidden fields' do
-      subject.collection_ids.should match /^<input/
-      subject.collection_ids.should match /name="ids\[\]"/
-      subject.collection_ids.should match /value="#{employee.id}"/
+      subject.collection_ids.should eq %|<input type="hidden" name="ids" value="#{employee_1.id}~#{employee_2.id}" />|
     end
 
     it "does not raise when the object is not a collection object" do
-      arguments[:form_object] = employee
+      arguments[:form_object] = employee_1
       expect{ subject.collection_ids }.not_to raise_error
     end
   end
@@ -30,9 +30,9 @@ describe ActionView::Helpers::FormBuilder do
       expect( subject ).to receive(:input).and_return "<simple-form-content>Simple Form Content</simple-form-content>".html_safe
       html = subject.optional_input(:section)
       doc = Nokogiri::HTML(html)
-      doc.css("[name='ids[]'][value='#{employee.id}']").should be_present # add ids if not yet set
+      doc.css("[name='ids'][value='#{employee_1.id}~#{employee_2.id}']").should be_present # add ids if not yet set
 
-      optional_div = doc.css("div.optional-input.optional-attribute-container.section.active.one[data-attribute='section'][data-one=true]").first
+      optional_div = doc.css("div.optional-input.optional-attribute-container.section.active[data-attribute='section'][data-one=false]").first
       optional_div.text.should eq 'Simple Form Content'
     end
   end
@@ -40,28 +40,25 @@ describe ActionView::Helpers::FormBuilder do
   describe '#get_optional_classes' do
 
     describe 'active/inactive' do
-      it "does not include active when record has no value" do
-        employee.section = nil
+      it "does not include active when one record has no value and the other one has" do
+        employee_1.section = nil
         subject.get_optional_classes(:section).should_not include'active'
         subject.get_optional_classes(:section).should include'inactive'
       end
 
-      it "includes active when record has en empty value" do
-        employee.section = ''
+      it "includes active when collection has an empty value" do
+        subject.object.section = ''
         subject.get_optional_classes(:section).should include'active'
         subject.get_optional_classes(:section).should_not include'inactive'
       end
 
       it 'includes active when two records have the same value' do
-        employee_2 = Employee.create section: 'SE1'
-        arguments[:form_object] = Employee::Collection.new([employee, employee_2])
         subject.get_optional_classes(:section).should include'active'
         subject.get_optional_classes(:section).should_not include'inactive'
       end
 
       it 'does not include active when two records have different values' do
-        employee_2 = Employee.create section: 'SE2'
-        arguments[:form_object] = Employee::Collection.new([employee, employee_2])
+        employee_2.section = 'SE2' # = Employee.create section: 'SE2'
         subject.get_optional_classes(:section).should_not include'active'
         subject.get_optional_classes(:section).should include'inactive'
       end
@@ -69,17 +66,16 @@ describe ActionView::Helpers::FormBuilder do
 
     describe 'one' do
       it "includes one if only one record is present" do
+        arguments[:form_object] = collection_class.new([employee_1])
         subject.get_optional_classes(:section).should include'one'
       end
 
       it "includes one if the form record is not a collection but a normal record" do
-        arguments[:form_object] = employee
+        arguments[:form_object] = employee_1
         subject.get_optional_classes(:section).should include'one'
       end
 
       it "does not include one for a collection having more than one records" do
-        employee_2 = Employee.create section: 'SE2'
-        arguments[:form_object] = Employee::Collection.new([employee, employee_2])
         subject.get_optional_classes(:section).should_not include'one'
       end
     end